Move clock function out of Callable and into Function
