+++ src/function.rs
@@ -1,4 +1,4 @@
-use std::fmt::Display;
+use std::{fmt::Display, time::{SystemTime, UNIX_EPOCH}};
 
 use crate::{
     callable::Call,
@@ -60,5 +60,45 @@ impl Call for Function {
 impl Display for Function {
     fn fmt(&self, f: &mut std::fmt::Formatter<'_>) -> std::fmt::Result {
         write!(f, "<fn {}>", self.name.lexeme())
+    }
+}
+
+#[derive(Debug, Clone, Default)]
+pub struct Clock {
+    arity: usize,
+}
+impl Clock {
+    pub fn new() -> Self {
+        Clock { arity: 0 }
+    }
+}
+
+impl Call for Clock {
+    fn call(
+        &mut self,
+        _interpreter: &mut Interpreter,
+        _arguments: Vec<Value>,
+    ) -> Result<Value, LoxError> {
+        let time = SystemTime::now()
+            .duration_since(UNIX_EPOCH)
+            .expect("Failed to get the system time")
+            .as_millis() as f64
+            / 1000.0;
+
+        Ok(Value::Number(time))
+    }
+
+    fn arity(&self) -> usize {
+        self.arity
+    }
+
+    fn box_clone(&self) -> Box<dyn Call> {
+        Box::new(self.clone())
+    }
+}
+
+impl Display for Clock {
+    fn fmt(&self, f: &mut std::fmt::Formatter<'_>) -> std::fmt::Result {
+        write!(f, "<native fn>")
     }
 }
