+++ src/parser.rs
@@ -69,6 +69,7 @@ impl Parser {
     fn function<S: Into<String>>(&mut self, kind: S) -> Result<Stmt, LoxError> {
         let kind = kind.into();
         let name = self.consume(TokenType::Identifier, format!("Expect {} name.", kind))?;
+        
         self.consume(TokenType::LeftParen, format!("Expect '(' after {} name.", kind))?;
         let mut parameters = vec![];
         if !self.check(TokenType::RightParen) {
